Add a pre and post parser hook to the interactive parser (#46)
* add an optional pre and post parser hook and and add_parser_arg hook

* update arrivial plotting

* fix font size and add pre/post/add_parser_args tests

* fix type conversion error

* cleanup testing error

* more scale factor cleanup

* fix more numpy min max type errors

* fix min max type casting

* revert changes

* alternative min max calls for dumps

* prevent agressive numpy data type casting

* more numpy1-numpy2 min/max fixes

* add a diagnostic print

* fix ambiguous 1d/2d/3d dump parsing

---------

+++ tests/my_test_opppy_dump_parser.py
@@ -32,7 +32,7 @@ class my_test_opppy_dump_parser():
         for line in lines:
             for key in keys:
                 if key in line and len(key) is len(line.split(':')[0]):
-                    data[key] = array(str_vector_to_float_vector(line.strip('\n').split(' ')[1:]))
+                    data[key] = array(str_vector_to_float_vector(line.strip('\n').split(' ')[1:])).astype(float)
 
 
         # build xy_verts for 2d mesh plotting example
@@ -44,4 +44,5 @@ class my_test_opppy_dump_parser():
 
         return data
         
+
